Add path constants + update links
