+++ web/screens/Entry/ModeChoice/index.tsx
@@ -4,18 +4,24 @@ import React from 'react'
 
 import Link from 'next/link'
 
+import {
+  onlineModeURL,
+  privateModePath,
+  privateModeURL,
+} from 'newcoin/constants/pathsToApp'
+
 import { useLoadURL } from '@/hooks/useLoadURL'
 
 import ChoiceItem from '@/screens/Entry/ModeChoice/ChoiceItem'
 
 const choicesList = [
   [
-    '/jan.html', //Change to '/jan' if you're in development mode
+    privateModePath, //Change to { privateModeURL } if you're in development mode
     'Private Mode',
     'Run the OS on your local computer. This mode is not connected to the Internet, which means all the data is private.',
   ],
   [
-    'https://os.newcoin.org',
+    onlineModeURL,
     'Online Mode',
     'Data will be stored on a remote server. You still have control over who can access it and you can decide to monetize your data.',
   ],
@@ -27,9 +33,14 @@ const EntryChoiceScreen = () => {
   return (
     <main className="relative min-h-screen bg-white">
       <div className="flex h-screen w-screen gap-x-6 p-6">
-        <Link href={choicesList[0][0]} className="block h-full w-full">
+        <div
+          onClick={() => {
+            onClickMode(choicesList[0][0])
+          }}
+          className="block h-full w-full"
+        >
           <ChoiceItem name={choicesList[0][1]} text={choicesList[0][2]} />
-        </Link>
+        </div>
         <div
           onClick={() => {
             onClickMode(choicesList[1][0])
